+++ main/src/main/java/net/oneandone/stool/cli/Stop.java
@@ -21,9 +21,14 @@ import net.oneandone.stool.util.Session;
 
 public class Stop extends StageCommand {
     private boolean sleep;
+    private int timeout;
 
     public Stop(Session session, boolean sleep) {
+        this(session, -1, sleep);
+    }
+    public Stop(Session session, int timeout, boolean sleep) {
         super(false, false, session, Mode.SHARED, Mode.SHARED, Mode.NONE);
+        this.timeout = timeout;
         this.sleep = sleep;
     }
 
@@ -41,7 +46,7 @@ public class Stop extends StageCommand {
                 console.info.println("going from sleeping to stopped.");
             }
         } else {
-            stage.stop(console);
+            stage.stop(console, timeout);
         }
         if (sleep) {
             if (!alreadySleeping) {